Move types => types/indexTypes
